requirements.txt
# requirements.txt
plotly
python-dotenv
requests
pandas
streamlit
statsmodels
prophet    # optional
watchdog   # optional
python-dotenv
